requirements.txt
anthropic>=0.35.0
google-genai>=0.8.0
pdfplumber>=0.11.0
pytesseract>=0.3.10
Pillow>=10.4.0
cairosvg>=2.7.1
selenium>=4.25.0
webdriver-manager>=4.0.2
beautifulsoup4>=4.12.3
requests>=2.32.3
python-dateutil>=2.9.0
reportlab>=4.2.5
python-dotenv>=1.0.1
lxml>=5.3.0
streamlit>=1.35.0
pandas>=2.0.0
pytest>=8.0.0
